First commit
earthranger_utils is a module to aide in the utilization of
a software called Earthranger to post various sensor data in a
way that will be useful for conservancy management, research that
requies data logging from remote sensors, and more. The first few
scripts are utilized by a package called CougarVision that posts
sorted camera trap images and logs the animals of interest a
camera has seen and when to Earthranger.